fix: separate integration tests and track combination fixture
