+++ test/fixtures/combination-component/index.js
@@ -1,0 +1,3 @@
+import CombinationComponentContainer from './components/CombinationComponentContainer';
+
+export default CombinationComponentContainer;
